Add/change DAOs and services for revised disease outbreak schema. Add tests. Change Common beans to be autowired by constructor. Allow suppression of checkstyle rules by using comments in the code. Note that diseasegroup lookup data is subject to change.

+++ src/Common/src/uk/ac/ox/zoo/seeg/abraid/mp/common/dao/HealthMapDiseaseDao.java
@@ -21,7 +21,7 @@ public interface HealthMapDiseaseDao {
      * @param id The ID.
      * @return The HealthMap disease, or null if not found.
      */
-    HealthMapDisease getById(Integer id);
+    HealthMapDisease getById(Long id);
 
     /**
      * Saves the specified HealthMap disease.
